Rename file to remove reference to ab testing
