+++ static/src/javascripts/projects/commercial/modules/cmp/cmp.spec.js
@@ -4,7 +4,7 @@ import fetchJson from 'lib/fetch-json';
 import { _, init } from './cmp';
 import { log as log_ } from './log';
 
-jest.mock('projects/commercial/modules/cmp/ccpa-ab-test', () => ({
+jest.mock('projects/commercial/modules/cmp/ccpa-cmp', () => ({
     isCcpaApplicable: () => false,
 }));
