+++ static/src/javascripts/projects/commercial/modules/cmp/stub.js
@@ -1,7 +1,7 @@
 // @flow
 /* eslint-disable no-underscore-dangle */
 import config from 'lib/config';
-import { isCcpaApplicable } from 'projects/commercial/modules/cmp/ccpa-ab-test';
+import { isCcpaApplicable } from 'commercial/modules/cmp/ccpa-cmp';
 
 if (config.get('switches.enableConsentManagementService') && !isCcpaApplicable()) {
     try {
